# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Era Patient - Hospital Patient Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      parameters:
        - name: stage
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Created patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Not found
    patch:
      operationId: updatePatient
      tags: [patients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientUpdate"
      responses:
        "200":
          description: Updated patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    delete:
      operationId: deletePatient
      tags: [patients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /patients/{id}/history:
    get:
      operationId: getPatientHistory
      tags: [patients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient full history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientHistory"
        "404":
          description: Not found
  /patients/{id}/checkin:
    post:
      operationId: checkinPatient
      tags: [patients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}/dequeue:
    post:
      operationId: dequeuePatient
      tags: [patients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient dequeued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}/treatment-plan:
    post:
      operationId: logTreatmentPlan
      tags: [patients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TreatmentPlanInput"
      responses:
        "200":
          description: Treatment plan logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}/flag-missed:
    post:
      operationId: flagMissedTreatment
      tags: [patients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlagMissedInput"
      responses:
        "200":
          description: Patient flagged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallTask"
...
  /wellness:
    get:
      operationId: listWellnessNewsletters
      tags: [wellness]
      responses:
        "200":
          description: All newsletters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WellnessNewsletter"
    put:
      operationId: upsertWellnessNewsletter
      tags: [wellness]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WellnessNewsletterInput"
      responses:
        "200":
          description: Saved newsletter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WellnessNewsletter"
...
  /wellness/{id}/send:
    post:
      operationId: markNewsletterSent
      tags: [wellness]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Newsletter marked as sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WellnessNewsletter"
components:
  schemas:
...
    Patient:
      type: object
      required: [id, firstName, lastName, stage, email, phone, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        hospitalId:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: string
        whatsappNumber:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        gender:
          type: ["string", "null"]
        stage:
          type: string
        preQueueStage:
          type: ["string", "null"]
        diagnosis:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        nextAppointment:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        treatmentPlan:
          type: ["string", "null"]
        treatmentType:
          type: ["string", "null"]
        medicationTiming:
          type: ["string", "null"]
        treatmentDurationDays:
          type: ["integer", "null"]
        treatmentEndDate:
          type: ["string", "null"]
        checkedInAt:
          type: ["string", "null"]
        treatmentStartedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    PatientInput:
      type: object
      required: [firstName, lastName, email, phone]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
        hospitalId:
          type: string
        email:
          type: string
        phone:
          type: string
        whatsappNumber:
          type: string
        age:
          type: integer
        gender:
          type: string
        stage:
          type: string
        diagnosis:
          type: string
        department:
          type: string
        nextAppointment:
          type: string
        notes:
          type: string
    PatientUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
        hospitalId:
          type: string
        email:
          type: string
        phone:
          type: string
        whatsappNumber:
          type: string
        age:
          type: integer
        gender:
          type: string
        stage:
          type: string
        diagnosis:
          type: string
        department:
          type: string
        nextAppointment:
          type: string
        notes:
          type: string
    PatientHistory:
      type: object
      required: [patient, activity, appointments, callTasks]
      properties:
        patient:
          $ref: "#/components/schemas/Patient"
        activity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        callTasks:
          type: array
          items:
            $ref: "#/components/schemas/CallTask"
    TreatmentPlanInput:
      type: object
      required: [treatmentPlan, treatmentType, treatmentDurationDays]
      properties:
        treatmentPlan:
          type: string
        treatmentType:
          type: string
        medicationTiming:
          type: string
        treatmentDurationDays:
          type: integer
        diagnosis:
          type: string
        department:
          type: string
    FlagMissedInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
...
    CallTask:
      type: object
      required: [id, patientId, patientName, phone, reason, flaggedAt, actionType]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        phone:
          type: string
        whatsappNumber:
          type: ["string", "null"]
        reason:
          type: string
        actionType:
          type: string
          description: "automated_message | manual_text | manual_call"
        outcome:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        flaggedAt:
          type: string
...
    Appointment:
      type: object
      required: [id, patientId, patientName, title, scheduledAt, status]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        title:
          type: string
        scheduledAt:
          type: string
        duration:
          type: integer
        department:
          type: ["string", "null"]
        status:
          type: string
...
    ActivityItem:
      type: object
      required: [id, type, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        patientId:
          type: ["integer", "null"]
        patientName:
          type: ["string", "null"]
        createdAt:
          type: string
        metadata:
          type: ["string", "null"]
...
    WellnessNewsletter:
      type: object
      required: [id, weekOf, content, updatedAt]
      properties:
        id:
          type: integer
        weekOf:
          type: string
        content:
          type: string
        lastSentAt:
          type: ["string", "null"]
        updatedAt:
          type: string
    WellnessNewsletterInput:
      type: object
      required: [content, weekOf]
      properties:
        content:
          type: string
        weekOf:
          type: string