# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Era Patient - Hospital Patient Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: number;
  firstName: string;
  lastName: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  hospitalId?: string | null;
  email: string;
  phone: string;
  /** @nullable */
  whatsappNumber?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  gender?: string | null;
  stage: string;
  /** @nullable */
  preQueueStage?: string | null;
  /** @nullable */
  diagnosis?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  nextAppointment?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  treatmentPlan?: string | null;
  /** @nullable */
  treatmentType?: string | null;
  /** @nullable */
  medicationTiming?: string | null;
  /** @nullable */
  treatmentDurationDays?: number | null;
  /** @nullable */
  treatmentEndDate?: string | null;
  /** @nullable */
  checkedInAt?: string | null;
  /** @nullable */
  treatmentStartedAt?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}