# lib/api-zod/src/generated/types/patientUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Era Patient - Hospital Patient Management API
 * OpenAPI spec version: 0.1.0
 */

export interface PatientUpdate {
  firstName?: string;
  lastName?: string;
  dateOfBirth?: string;
  hospitalId?: string;
  email?: string;
  phone?: string;
  whatsappNumber?: string;
  age?: number;
  gender?: string;
  stage?: string;
  diagnosis?: string;
  department?: string;
  nextAppointment?: string;
  notes?: string;
}