# lib/api-zod/src/generated/types/wellnessNewsletterInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Era Patient - Hospital Patient Management API
 * OpenAPI spec version: 0.1.0
 */

export interface WellnessNewsletterInput {
  content: string;
  weekOf: string;
}